Added sound for the Scrapper, Lancer, and Sentinel. Fixed some conversion warnings across the engine. Attempted to fix SoundSpace::Clear. Temporarily turned off some physics debug output. Updated the logic in ball controller so that the impact sound can't be spammed. Added logic to the AntiEngeryField zilch script. Updated the player controller with slightly better logic for audio movement.

+++ DCEngine/Core/Components/Orientation.cpp
@@ -70,7 +70,7 @@ namespace DCEngine
 
       // convert to degrees first
 
-      DeltaRot = DeltaRot * 180.0f / 3.14159265359;
+      DeltaRot = static_cast<float>(DeltaRot * 180.0f / 3.14159265359);
 
       this->Owner()->getComponent<Components::Transform>()->Rotation.z -= DeltaRot;
